+++ packages/mirai/lib/src/parsers/mirai_alert_dialog/mirai_alert_dialog.dart
@@ -31,7 +31,7 @@ class MiraiAlertDialog with _$MiraiAlertDialog {
     double? elevation,
     String? semanticLabel,
     @Default(MiraiEdgeInsets(left: 40, right: 40, top: 24, bottom: 24))
-        MiraiEdgeInsets insetPadding,
+    MiraiEdgeInsets insetPadding,
     @Default(Clip.none) Clip clipBehavior,
     @Default(false) bool scrollable,
   }) = _MiraiAlertDialog;