+++ packages/mirai/lib/src/parsers/mirai_card_theme_data/mirai_card_theme_data.freezed.dart
@@ -1,0 +1,313 @@
+// coverage:ignore-file
+// GENERATED CODE - DO NOT MODIFY BY HAND
+// ignore_for_file: type=lint
+// ignore_for_file: unused_element, deprecated_member_use, deprecated_member_use_from_same_package, use_function_type_syntax_for_parameters, unnecessary_const, avoid_init_to_null, invalid_override_different_default_values_named, prefer_expression_function_bodies, annotate_overrides, invalid_annotation_target, unnecessary_question_mark
+
+part of 'mirai_card_theme_data.dart';
+
+// **************************************************************************
+// FreezedGenerator
+// **************************************************************************
+
+T _$identity<T>(T value) => value;
+
+final _privateConstructorUsedError = UnsupportedError(
+    'It seems like you constructed your class using `MyClass._()`. This constructor is only meant to be used by freezed and you are not supposed to need it nor use it.\nPlease check the documentation here for more information: https://github.com/rrousselGit/freezed#custom-getters-and-methods');
+
+MiraiCardThemeData _$MiraiCardThemeDataFromJson(Map<String, dynamic> json) {
+  return _MiraiCardThemeData.fromJson(json);
+}
+
+/// @nodoc
+mixin _$MiraiCardThemeData {
+  Clip? get clipBehavior => throw _privateConstructorUsedError;
+  String? get color => throw _privateConstructorUsedError;
+  String? get shadowColor => throw _privateConstructorUsedError;
+  String? get surfaceTintColor => throw _privateConstructorUsedError;
+  double? get elevation => throw _privateConstructorUsedError;
+  MiraiEdgeInsets? get margin => throw _privateConstructorUsedError;
+  MiraiBorder? get shape => throw _privateConstructorUsedError;
+
+  Map<String, dynamic> toJson() => throw _privateConstructorUsedError;
+  @JsonKey(ignore: true)
+  $MiraiCardThemeDataCopyWith<MiraiCardThemeData> get copyWith =>
+      throw _privateConstructorUsedError;
+}
+
+/// @nodoc
+abstract class $MiraiCardThemeDataCopyWith<$Res> {
+  factory $MiraiCardThemeDataCopyWith(
+          MiraiCardThemeData value, $Res Function(MiraiCardThemeData) then) =
+      _$MiraiCardThemeDataCopyWithImpl<$Res, MiraiCardThemeData>;
+  @useResult
+  $Res call(
+      {Clip? clipBehavior,
+      String? color,
+      String? shadowColor,
+      String? surfaceTintColor,
+      double? elevation,
+      MiraiEdgeInsets? margin,
+      MiraiBorder? shape});
+
+  $MiraiEdgeInsetsCopyWith<$Res>? get margin;
+  $MiraiBorderCopyWith<$Res>? get shape;
+}
+
+/// @nodoc
+class _$MiraiCardThemeDataCopyWithImpl<$Res, $Val extends MiraiCardThemeData>
+    implements $MiraiCardThemeDataCopyWith<$Res> {
+  _$MiraiCardThemeDataCopyWithImpl(this._value, this._then);
+
+  // ignore: unused_field
+  final $Val _value;
+  // ignore: unused_field
+  final $Res Function($Val) _then;
+
+  @pragma('vm:prefer-inline')
+  @override
+  $Res call({
+    Object? clipBehavior = freezed,
+    Object? color = freezed,
+    Object? shadowColor = freezed,
+    Object? surfaceTintColor = freezed,
+    Object? elevation = freezed,
+    Object? margin = freezed,
+    Object? shape = freezed,
+  }) {
+    return _then(_value.copyWith(
+      clipBehavior: freezed == clipBehavior
+          ? _value.clipBehavior
+          : clipBehavior // ignore: cast_nullable_to_non_nullable
+              as Clip?,
+      color: freezed == color
+          ? _value.color
+          : color // ignore: cast_nullable_to_non_nullable
+              as String?,
+      shadowColor: freezed == shadowColor
+          ? _value.shadowColor
+          : shadowColor // ignore: cast_nullable_to_non_nullable
+              as String?,
+      surfaceTintColor: freezed == surfaceTintColor
+          ? _value.surfaceTintColor
+          : surfaceTintColor // ignore: cast_nullable_to_non_nullable
+              as String?,
+      elevation: freezed == elevation
+          ? _value.elevation
+          : elevation // ignore: cast_nullable_to_non_nullable
+              as double?,
+      margin: freezed == margin
+          ? _value.margin
+          : margin // ignore: cast_nullable_to_non_nullable
+              as MiraiEdgeInsets?,
+      shape: freezed == shape
+          ? _value.shape
+          : shape // ignore: cast_nullable_to_non_nullable
+              as MiraiBorder?,
+    ) as $Val);
+  }
+
+  @override
+  @pragma('vm:prefer-inline')
+  $MiraiEdgeInsetsCopyWith<$Res>? get margin {
+    if (_value.margin == null) {
+      return null;
+    }
+
+    return $MiraiEdgeInsetsCopyWith<$Res>(_value.margin!, (value) {
+      return _then(_value.copyWith(margin: value) as $Val);
+    });
+  }
+
+  @override
+  @pragma('vm:prefer-inline')
+  $MiraiBorderCopyWith<$Res>? get shape {
+    if (_value.shape == null) {
+      return null;
+    }
+
+    return $MiraiBorderCopyWith<$Res>(_value.shape!, (value) {
+      return _then(_value.copyWith(shape: value) as $Val);
+    });
+  }
+}
+
+/// @nodoc
+abstract class _$$_MiraiCardThemeDataCopyWith<$Res>
+    implements $MiraiCardThemeDataCopyWith<$Res> {
+  factory _$$_MiraiCardThemeDataCopyWith(_$_MiraiCardThemeData value,
+          $Res Function(_$_MiraiCardThemeData) then) =
+      __$$_MiraiCardThemeDataCopyWithImpl<$Res>;
+  @override
+  @useResult
+  $Res call(
+      {Clip? clipBehavior,
+      String? color,
+      String? shadowColor,
+      String? surfaceTintColor,
+      double? elevation,
+      MiraiEdgeInsets? margin,
+      MiraiBorder? shape});
+
+  @override
+  $MiraiEdgeInsetsCopyWith<$Res>? get margin;
+  @override
+  $MiraiBorderCopyWith<$Res>? get shape;
+}
+
+/// @nodoc
+class __$$_MiraiCardThemeDataCopyWithImpl<$Res>
+    extends _$MiraiCardThemeDataCopyWithImpl<$Res, _$_MiraiCardThemeData>
+    implements _$$_MiraiCardThemeDataCopyWith<$Res> {
+  __$$_MiraiCardThemeDataCopyWithImpl(
+      _$_MiraiCardThemeData _value, $Res Function(_$_MiraiCardThemeData) _then)
+      : super(_value, _then);
+
+  @pragma('vm:prefer-inline')
+  @override
+  $Res call({
+    Object? clipBehavior = freezed,
+    Object? color = freezed,
+    Object? shadowColor = freezed,
+    Object? surfaceTintColor = freezed,
+    Object? elevation = freezed,
+    Object? margin = freezed,
+    Object? shape = freezed,
+  }) {
+    return _then(_$_MiraiCardThemeData(
+      clipBehavior: freezed == clipBehavior
+          ? _value.clipBehavior
+          : clipBehavior // ignore: cast_nullable_to_non_nullable
+              as Clip?,
+      color: freezed == color
+          ? _value.color
+          : color // ignore: cast_nullable_to_non_nullable
+              as String?,
+      shadowColor: freezed == shadowColor
+          ? _value.shadowColor
+          : shadowColor // ignore: cast_nullable_to_non_nullable
+              as String?,
+      surfaceTintColor: freezed == surfaceTintColor
+          ? _value.surfaceTintColor
+          : surfaceTintColor // ignore: cast_nullable_to_non_nullable
+              as String?,
+      elevation: freezed == elevation
+          ? _value.elevation
+          : elevation // ignore: cast_nullable_to_non_nullable
+              as double?,
+      margin: freezed == margin
+          ? _value.margin
+          : margin // ignore: cast_nullable_to_non_nullable
+              as MiraiEdgeInsets?,
+      shape: freezed == shape
+          ? _value.shape
+          : shape // ignore: cast_nullable_to_non_nullable
+              as MiraiBorder?,
+    ));
+  }
+}
+
+/// @nodoc
+@JsonSerializable()
+class _$_MiraiCardThemeData implements _MiraiCardThemeData {
+  const _$_MiraiCardThemeData(
+      {this.clipBehavior,
+      this.color,
+      this.shadowColor,
+      this.surfaceTintColor,
+      this.elevation,
+      this.margin,
+      this.shape});
+
+  factory _$_MiraiCardThemeData.fromJson(Map<String, dynamic> json) =>
+      _$$_MiraiCardThemeDataFromJson(json);
+
+  @override
+  final Clip? clipBehavior;
+  @override
+  final String? color;
+  @override
+  final String? shadowColor;
+  @override
+  final String? surfaceTintColor;
+  @override
+  final double? elevation;
+  @override
+  final MiraiEdgeInsets? margin;
+  @override
+  final MiraiBorder? shape;
+
+  @override
+  String toString() {
+    return 'MiraiCardThemeData(clipBehavior: $clipBehavior, color: $color, shadowColor: $shadowColor, surfaceTintColor: $surfaceTintColor, elevation: $elevation, margin: $margin, shape: $shape)';
+  }
+
+  @override
+  bool operator ==(dynamic other) {
+    return identical(this, other) ||
+        (other.runtimeType == runtimeType &&
+            other is _$_MiraiCardThemeData &&
+            (identical(other.clipBehavior, clipBehavior) ||
+                other.clipBehavior == clipBehavior) &&
+            (identical(other.color, color) || other.color == color) &&
+            (identical(other.shadowColor, shadowColor) ||
+                other.shadowColor == shadowColor) &&
+            (identical(other.surfaceTintColor, surfaceTintColor) ||
+                other.surfaceTintColor == surfaceTintColor) &&
+            (identical(other.elevation, elevation) ||
+                other.elevation == elevation) &&
+            (identical(other.margin, margin) || other.margin == margin) &&
+            (identical(other.shape, shape) || other.shape == shape));
+  }
+
+  @JsonKey(ignore: true)
+  @override
+  int get hashCode => Object.hash(runtimeType, clipBehavior, color, shadowColor,
+      surfaceTintColor, elevation, margin, shape);
+
+  @JsonKey(ignore: true)
+  @override
+  @pragma('vm:prefer-inline')
+  _$$_MiraiCardThemeDataCopyWith<_$_MiraiCardThemeData> get copyWith =>
+      __$$_MiraiCardThemeDataCopyWithImpl<_$_MiraiCardThemeData>(
+          this, _$identity);
+
+  @override
+  Map<String, dynamic> toJson() {
+    return _$$_MiraiCardThemeDataToJson(
+      this,
+    );
+  }
+}
+
+abstract class _MiraiCardThemeData implements MiraiCardThemeData {
+  const factory _MiraiCardThemeData(
+      {final Clip? clipBehavior,
+      final String? color,
+      final String? shadowColor,
+      final String? surfaceTintColor,
+      final double? elevation,
+      final MiraiEdgeInsets? margin,
+      final MiraiBorder? shape}) = _$_MiraiCardThemeData;
+
+  factory _MiraiCardThemeData.fromJson(Map<String, dynamic> json) =
+      _$_MiraiCardThemeData.fromJson;
+
+  @override
+  Clip? get clipBehavior;
+  @override
+  String? get color;
+  @override
+  String? get shadowColor;
+  @override
+  String? get surfaceTintColor;
+  @override
+  double? get elevation;
+  @override
+  MiraiEdgeInsets? get margin;
+  @override
+  MiraiBorder? get shape;
+  @override
+  @JsonKey(ignore: true)
+  _$$_MiraiCardThemeDataCopyWith<_$_MiraiCardThemeData> get copyWith =>
+      throw _privateConstructorUsedError;
+}