+++ packages/mirai/lib/src/parsers/mirai_visual_density/mirai_visual_density.g.dart
@@ -1,0 +1,21 @@
+// GENERATED CODE - DO NOT MODIFY BY HAND
+
+part of 'mirai_visual_density.dart';
+
+// **************************************************************************
+// JsonSerializableGenerator
+// **************************************************************************
+
+_$_MiraiVisualDensity _$$_MiraiVisualDensityFromJson(
+        Map<String, dynamic> json) =>
+    _$_MiraiVisualDensity(
+      horizontal: (json['horizontal'] as num).toDouble(),
+      vertical: (json['vertical'] as num).toDouble(),
+    );
+
+Map<String, dynamic> _$$_MiraiVisualDensityToJson(
+        _$_MiraiVisualDensity instance) =>
+    <String, dynamic>{
+      'horizontal': instance.horizontal,
+      'vertical': instance.vertical,
+    };