+++ packages/mirai/CHANGELOG.md
@@ -1,3 +1,33 @@
+## 0.3.0
+
+- Added support for the appBarTheme (#96) by @i-asimkhan
+- Added support for the textButtonTheme (#98) by @i-asimkhan
+- Added support for the floatingActionButtonTheme (#99) by @i-asimkhan
+- Added support for the bottomNavBarTheme (#136) by @i-asimkhan
+- Added support for the bottomSheetTheme (#137) by @i-asimkhan
+- Added support for the CardTheme (#139) by @i-asimkhan
+- Added support for the dialog theme (#140) by @i-asimkhan
+- Added support for the MiraiNavigationBarThemeData (#142) by @i-asimkhan
+- Added support for the MiraiTabBarThemeData (#143) by @i-asimkhan
+- Added support for the MiraiTabBarThemeData (#143) by @i-asimkhan
+- Added CODE_OF_CONDUCT(#122) by @divyanshub024
+
+## 0.2.0
+
+* BREAKING: feat: Call `Mirai.initialize` to initialize Mirai (#103)
+* feat: Add support for iconTheme (#91)
+* feat: Add FractionallySizedBox widget (#100)
+* feat: Add support for elevatedButtonTheme (#92)
+* feat: Add support for outlinedButtonTheme (#93)
+* feat: Add support for iconButtonTheme (#94)
+* feat: Add center widget (#102)
+* feat: Add support to inject dio instance (#104)
+* feat: Add support for InputFormatters (#111)
+* feat: Add support for InputValidation (#106)
+* feat: Add `BoxDecoration` support for Container (#112)
+* feat: Add Stack and Positioned widget (#114)
+* chore: Update Readme (#101)
+
 ## 0.1.0
 
 * Added Mirai Theme support